# Documentos/api.yaml
openapi: 3.0.3
info:
  title: API Sistema de Entregas por Paracaídas
  version: 1.1.0
  description: Contrato REST para el sistema de encomiendas de caída libre. Autenticación delegada a Clerk (ADR-006).
servers:
  - url: http://localhost:3000/api/v1
    description: Servidor de desarrollo

tags:
  - name: Solicitudes
    description: Gestión de pedidos de encomiendas y su historial
  - name: Bases
    description: Gestión de bases y stock de los remitentes
  - name: Envios
    description: Gestión de envíos y contenedores
  - name: Usuarios
    description: Gestión de cuentas y roles

# Todas las rutas requieren el token de Clerk por defecto
security:
  - clerkAuth: []

paths:
  # ==========================================
  # MÓDULO: USUARIOS
  # ==========================================
  /usuarios:
    get:
      summary: Listar usuarios del sistema
      tags: [Usuarios]
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Usuario'

  /usuarios/{id}/rol:
    patch:
      summary: Modificar el rol de un usuario (Solo Admin)
      tags: [Usuarios]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rol:
                  type: string
                  enum: [solicitante, remitente, admin]
      responses:
        '200':
          description: Rol actualizado
        '403':
          description: No autorizado

  # ==========================================
  # MÓDULO: SOLICITUDES E HISTORIAL
  # ==========================================
  /solicitudes:
    post:
      summary: Crear una nueva solicitud de encomienda
      tags: [Solicitudes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NuevaSolicitudRequest'
      responses:
        '201':
          description: Solicitud creada (el estado indica si fue aprobada o rechazada por stock)
          headers:
            Location:
              description: URL de la nueva solicitud
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CrearSolicitudResponse'

    get:
      summary: Listar solicitudes (filtrables por estado)
      tags: [Solicitudes]
      parameters:
        - name: estado
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Lista de solicitudes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Solicitud'

  /solicitudes/{id}:
    get:
      summary: Obtener el detalle de una solicitud específica
      tags: [Solicitudes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalle de la solicitud
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Solicitud'
        '404':
          description: Solicitud no encontrada

  /solicitudes/{id}/estado:
    patch:
      summary: Actualizar el estado de una solicitud
      tags: [Solicitudes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nuevoEstado]
              properties:
                nuevoEstado:
                  type: string
                  example: En preparación
      responses:
        '200':
          description: Estado actualizado exitosamente
        '422':
          description: Transición de estado inválida

  /solicitudes/{id}/historial:
    get:
      summary: Obtener el historial de cambios de estado (RNF16)
      tags: [Solicitudes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Historial de la solicitud
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistorialEstado'

  # ==========================================
  # MÓDULO: BASES Y STOCK
  # ==========================================
  /bases/{id}/stock:
    get:
      summary: Consultar inventario disponible en una base
      tags: [Bases]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de stock de la base
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockBase'

    put:
      summary: Actualizar inventario físico de la base
      tags: [Bases]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  productoId:
                    type: string
                  cantidadAAgregar:
                    type: integer
      responses:
        '200':
          description: Stock actualizado

  # ==========================================
  # MÓDULO: ENVÍOS Y CONTENEDORES
  # ==========================================
  /envios:
    post:
      summary: Programar un nuevo envío
      tags: [Envios]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Envio'
      responses:
        '201':
          description: Envío creado
    get:
      summary: Listar envíos programados o en curso
      tags: [Envios]
      responses:
        '200':
          description: Lista de envíos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Envio'

  /envios/{id}/contenedores:
    post:
      summary: Asignar un contenedor (paquete físico con paracaídas) a un envío
      tags: [Envios]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Contenedor'
      responses:
        '201':
          description: Contenedor asignado al envío

components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT provisto por Clerk en el header Authorization

  schemas:
    Usuario:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
        email:
          type: string
        rol:
          type: string
        baseId:
          type: integer
          nullable: true

    NuevaSolicitudRequest:
      type: object
      required: [latDestino, lonDestino, prioridad, productos]
      properties:
        latDestino:
          type: number
          format: float
          example: -38.7183
        lonDestino:
          type: number
          format: float
          example: -62.2663
        prioridad:
          type: string
          enum: [Alta, Media, Baja]
        productos:
          type: array
          items:
            type: object
            required: [productoId, cantidad]
            properties:
              productoId:
                type: string
                format: uuid
              cantidad:
                type: integer
                minimum: 1

    Solicitud:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fechaSolicitada:
          type: string
          format: date-time
        estado:
          type: string
          example: Creada
        latDestino:
          type: number
          format: float
        lonDestino:
          type: number
          format: float
        detalles:
          type: array
          items:
            $ref: '#/components/schemas/DetalleSolicitud'

    DetalleSolicitud:
      type: object
      properties:
        productoId:
          type: string
          format: uuid
        cantidadSolicitada:
          type: integer

    HistorialEstado:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fechaHora:
          type: string
          format: date-time
        estadoAnterior:
          type: string
        estadoNuevo:
          type: string
        usuarioId:
          type: string

    StockBase:
      type: object
      properties:
        productoId:
          type: string
          format: uuid
        nombre:
          type: string
        cantidadDisponible:
          type: integer
        cantidadReservada:
          type: integer

    Envio:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fechaHoraProgramada:
          type: string
          format: date-time
        estado:
          type: string
          example: Programado

    Contenedor:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tipoParacaidas:
          type: string
        pesoMax:
          type: number
        envioId:
          type: string
          format: uuid

    CrearSolicitudResponse:
      type: object
      properties:
        solicitud:
          $ref: '#/components/schemas/Solicitud'
        asignada:
          type: boolean
          description: Indica si se pudo asignar un remitente automáticamente
        stockFaltante:
          type: array
          items:
            type: string
          description: IDs de productos sin stock suficiente (solo si asignada es false)

    Error:
      type: object
      properties:
        code:
          type: string
          example: TRANSICION_INVALIDA
        message:
          type: string
          example: No se puede cancelar una solicitud en estado "Lanzada".
        details:
          type: array
          items:
            type: string